Use return inside operator function
The code before was to complicated, when a first result is
calculated, then there are only two possibilties left, either "clear"
or adding an operator and a number to the result.
So all that had to be done is to set the first number to the result as
a string so that it can be converted to a number again, clear operator
and second number and and return. Then the result operator second
number can be calculated and so on